==> Lessons/05_Part1_DataVisualization.Rmd <==
---
title: "5: Part 1 - Data Visualization Basics"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
#install.packages('formatR')
knitr::opts_chunk$set(tidy.opts=list(width.cutoff=80), tidy=TRUE)
```

## Objectives
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## Opening discussion
Effective data visualization depends on purposeful choices about graph types. The ideal graph type depends on the type of data and the message the visualizer desires to communicate. The best visualizations are clear and simple. A good resource for data visualization is [Data to Viz](https://www.data-to-viz.com/), which includes both a decision tree for visualization types and explanation pages for each type of data, including links to R resources to create them. Take a few minutes to explore this website.

## Set Up

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse)
#install.packages("ggridges")
library(ggridges)

PeterPaul.chem.nutrients <- 
  read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv", stringsAsFactors = TRUE)
PeterPaul.chem.nutrients.gathered <-
  read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv", stringsAsFactors = TRUE)
EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")
PeterPaul.chem.nutrients.gathered$sampledate <- as.Date(PeterPaul.chem.nutrients.gathered$sampledate, format = "%Y-%m-%d")

```

## ggplot

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_density_ridges
* geom_point
* geom_errorbar
* geom_smooth
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape
* size
* transparency

### Plotting continuous variables over time: Scatterplot and Line Plot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
ggplot(EPAair, aes(x = Date, y = Ozone)) + 
  geom_point()

O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot)

# Fix this code
O3plot2 <- ggplot(EPAair) +
<<<<<<< HEAD
 # geom_point(aes(x = Date, y = Ozone, color = "blue"))
  geom_point(aes(x = Date, y = Ozone), color = "blue")
# keep color out of the aes() function
=======
  geom_point(aes(x = Date, y = Ozone, color = "blue"))
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
print(O3plot2)

# Add additional variables
# How could you automatically assign a marker color to a variable? 
PMplot <- 
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year), color = Site.Name)) +
  geom_point()
print(PMplot)

# Separate plot with facets
PMplot.faceted <-
  ggplot(EPAair, aes(x = Month, y = PM2.5, shape = as.factor(Year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3)
print(PMplot.faceted)

<<<<<<< HEAD
# this will create one plot for each site instead of just having the sites being in different colors
#as.factor(Year) makes it so the different variables are different shapes 

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Filter dataset within plot building and facet by multiple variables
PMplot.faceted2 <- 
  ggplot(subset(EPAair, Site.Name == "Clemmons Middle" | Site.Name == "Leggett" |
                  Site.Name == "Bryson City"), 
         aes(x = Month, y = PM2.5)) + 
  geom_point() +
  facet_grid(Site.Name ~ Year) 
print(PMplot.faceted2)

<<<<<<< HEAD
#this is if you only want some of the sites in your analysis and this plot also will separate the years into the columns with   facet_grid(Site.Name ~ Year) 

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Plot true time series with geom_line
PMplot.line <- 
  ggplot(subset(EPAair, Site.Name == "Leggett"), 
         aes(x = Date, y = PM2.5)) +
  geom_line()
print(PMplot.line)

<<<<<<< HEAD
# no longer geom_line() this will be a typical example of a time series plot 

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b

```


### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO)

<<<<<<< HEAD
#the two points are so far from the others that the rest are so concentrated together - so you need to adjust the axis, adjust them to exclude the outliers and zoom to see the majority of the data

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Adjust axes
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
<<<<<<< HEAD
  xlim(0, 2500) +
  ylim(0, 20)
print(lightvsDOfixed)


# know explore the relationship b/t depth and temp. if you dont specify scale_y_reverse() you will get it from 0-20, but for depth you want it from 20-0

=======
  xlim(0, 250) +
  ylim(0, 20)
print(lightvsDOfixed)

>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Depth in the fields of limnology and oceanography is on a reverse scale
tempvsdepth <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse()
print(tempvsdepth)

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() +
  geom_smooth(method = lm) +
  geom_abline(aes(slope = 16, intercept = 0))
print(NvsP)

<<<<<<< HEAD
# if you want to find a trend line (geom_smooth() - you need to chose a method (method-lm is linear))
# geom_abline() is if you already know the slope of the line

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
```

### Plotting continuous vs. categorical variables
A traditional way to display summary statistics of continuous variables is a bar plot with error bars. Let's explore why this might not be the most effective way to display this type of data. Navigate to the Caveats page on Data to Viz (https://www.data-to-viz.com/caveats.html) and find the page that explores barplots and error bars. 

What might be more effective ways to display the information? Navigate to the boxplots page in the Caveats section to explore further.

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"?
<<<<<<< HEAD
  #color regards the outline of the shape and fill will be the filling of the shape 
# when you are working with box plots always use color to change 
print(Nutrientplot3)

#bar plot will not have the best representation often because it will not always show different distributions 
#you can end up leaving things out from the dataset 
#you can sometimes use a boxplot with the points over it
# look at the data to vis website for more info 



=======
print(Nutrientplot3)

>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_dotplot(aes(color = nutrient, fill = nutrient), binaxis = "y", binwidth = 1, 
               stackdir = "center", position = "dodge", dotsize = 2) #
print(Nutrientplot4)

<<<<<<< HEAD
#search for where these arguments are located
# binaxis = "Y" is the stacking process of the observations over y 
#binwidth - sets with of bin
#stackdir = center is the stacking the dots in the center. 
#position = dodge is so that the points don't get stacked 

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Violin plot
Nutrientplot5 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #
print(Nutrientplot5)

<<<<<<< HEAD
# rotated density plot (violin plot)

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Frequency polygons
# Using a tidy dataset
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "darkred") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "blue") +
  geom_freqpoly(aes(x = no23), color = "royalblue") +
  geom_freqpoly(aes(x = po4), color = "red") 
print(Nutrientplot6)

<<<<<<< HEAD
# for each variable specify the variable wanted on the x axis and then color
# problems with this (no legend) - the x-axis is only adding the name of the first variable 
# this is not a good way of generating a frequency plot 

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Using a gathered dataset
Nutrientplot7 <-   
  ggplot(PeterPaul.chem.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)

<<<<<<< HEAD
# this is when the pivot_longer and pivot_wider (gathered) it can be easier to plot 
#compare these before doing the homework to see the differences

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
# Frequency polygons have the risk of becoming spaghetti plots. 
# See <https://www.data-to-viz.com/caveat/spaghetti.html> for more info.

# Ridgeline plot
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients.gathered, aes(y = nutrient, x = concentration)) +
  geom_density_ridges(aes(fill = lakename), alpha = 0.5) #
print(Nutrientplot6)

<<<<<<< HEAD
# if you have a lot of variables you run the risk of a frequency plot becoming a spaghetti plot 
#ridgeline will show the density over time 
# used fill to color the inside of the curves

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b

```
