==> Lessons/06_Lab_GLMs.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
<<<<<<< HEAD
install.packages("agricolae")

=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
library(tidyverse)
library(agricolae)   

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```



## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 1:
<<<<<<< HEAD
Density_plot <- ggplot(EPAair, aes(X=Ozone, fill = as.factor(Year))) + 
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("darkgreen", "darkblue")) +
  labs(x = "Ozone AQI value")
    print(Density_plot)

  boxplot <- 
    ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_boxplot()

  
  
  
# code for these are in my camera roll
=======


>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

<<<<<<< HEAD
```{r, fig.height = 3, fig.width = 4}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 

EPA_regression <- lm(data = EPAair, PM2.5 ~ Ozone)
summary(EPA_regression)

# we assume PM2.5 is our response (so y)
# a small r^2 value means that there are others things that are predictor for the variability in PM2.5 (not just ozone)
=======
```{r}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 

>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b


#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot

<<<<<<< HEAD
EPA_regression_plot <- 
  ggplot(EPAair, aes(x=Ozone, y=PM2.5)) +
    geom_point() +
  geom_smooth(method = "lm", col = "blue")
print(EPA_regression_plot)

#next step is trying to add more variables to model 

#check phone for pictures with more info
=======

>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b

```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.

<<<<<<< HEAD
install.packages("corrplot")
library(corrplot)
library(dplyr)

EPAair_predict <- 
  EPAair %>%
  select(Mean_latitude:Ozone) %>% 
  drop_na()

EPAairCorr <- cor(EPAair_predict)
corrplot(EPAairCorr, method = "ellipse")

#gives us a visual representation of the correlation - the better the correlation the straighter the shape of the ellipses 
#color - reds and oranges (negative correlation) + blues (postice correlation)

corrplot.mixed(EPAairCorr, upper = "ellipse")

#gives us a better indication of what we need to add to the linear regression 

#from this correlation plot, you can see that you should run the regression with the mean_latitude, month, and ozone(which we already know) 

#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 5?

EPAAIC <- lm(data = EPAair_predict, PM2.5 ~ Ozone + Year + Month + Mean_latitude + Mean_longitude)

step(EPAAIC)
# step tells us that if we remove year we will improve in terms of AIC (AIC without year is one unit less than none)

# it is important to consider more than one metric when choosing the model you are going to go with 

summary(EPAAIC)

#need to run summary
# with p-values we can make assumptions for variables 
# shows similar results to what we got with the corrplot() 

#results from step() show that ozone and month are correlated 
=======



#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 5?


>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b


#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 5. 

<<<<<<< HEAD
AICmodelEPA <- lm(data = EPAair_predict, PM2.5 ~ Ozone + Mean_latitude + Month + Mean_longitude)
summary(AICmodelEPA)

#take out year because it wasn't a good variable for our regression, so with the new variables we get a higher r-squared 

# the p-values in this regression show that there is a relationship between month and ozone with pm2.5 
=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
```



## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups

Litter.Totals <- Litter %>%
  mutate( treatgroups = Litter.Totals.groups$groups[plotID,2])

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot```

<<<<<<< HEAD

#a. order by groups - create the column groups in your dataset - order by alphabetical 



assign.group <- function(site)
  return(Litter.Totals.groups$groups[as.character(site),2])




# picture of code is in camera roll 
  


=======
>>>>>>> fc8c57328ec11d4acfebcceb37adf859baecba8b
